[Filestore] extract test helpers for service to separate file
