+++ cloud/filestore/libs/storage/service/service_actor_actions_ut.cpp
@@ -1,4 +1,5 @@
 #include "service.h"
+#include "service_ut_helpers.h"
 
 #include <cloud/filestore/libs/storage/testlib/service_client.h>
 #include <cloud/filestore/libs/storage/testlib/tablet_client.h>
@@ -71,15 +72,6 @@ NProtoPrivate::TChangeStorageConfigResponse ExecuteChangeStorageConfig(
     UNIT_ASSERT_C(status.ok(), ToString(status.message()));
 
     return response;
-}
-
-TString GenerateValidateData(ui32 size, ui32 seed = 0)
-{
-    TString data(size, 0);
-    for (ui32 i = 0; i < size; ++i) {
-        data[i] = 'A' + ((i + seed) % ('Z' - 'A' + 1));
-    }
-    return data;
 }
 
 void WaitForTabletStart(TServiceClient& service)
