+++ cloud/filestore/libs/storage/service/ut/ya.make
@@ -5,6 +5,7 @@ INCLUDE(${ARCADIA_ROOT}/cloud/filestore/tests/recipes/medium.inc)
 SRCS(
     helpers_ut.cpp
     service_ut.cpp
+    service_ut_helpers.cpp
     service_ut_parentless.cpp
     service_ut_sharding.cpp
     service_actor_actions_ut.cpp
